[bugfix] Only compare matching length vectors
Ensure `length(special.days) == 1L` before comparing to `"weekends"`.
This is requried for newer versions of R.

As far as I can tell, this does not limit any functionality, but masks
the informative error message when specifying the wrong number of days
(i.e. 365 for a leap year) with an uninformative one.

This commit fixes issue #30.

Signed-off-by: Marcel Schilling <foss@mschilli.com>